Not sure how to resolve the errors. Did what I could.
